+++ python/tests/unit/test_irauth.py
@@ -21,6 +21,13 @@ from ambassador.utils import NullSecretHandler
 from tests.utils import default_listener_manifests
 
 
+def _assert_ext_auth_disabled(route):
+    assert route
+    per_filter_config = route.get("typed_per_filter_config")
+    assert per_filter_config.get("envoy.filters.http.ext_authz")
+    assert per_filter_config.get("envoy.filters.http.ext_authz").get("disabled") == True
+
+
 def _get_ext_auth_config(yaml):
     for listener in yaml["static_resources"]["listeners"]:
         for filter_chain in listener["filter_chains"]:
@@ -200,9 +207,10 @@ spec:
 
 
 @pytest.mark.compilertest
-def test_basic_http_redirect_disables_ext_authz():
-    """Test that http --> https redirect route exists along with
-    typed_per_filter_config for disabling ext_authz when an AuthService exists
+def test_redirects_disables_ext_authz():
+    """Test that the ext_authz is disabled on envoy redirect routes
+    for https_redirects and host_redirects. This is to ensure that the
+    redirect occurs before making any calls to the ext_authz service
     """
 
     if EDGE_STACK:
@@ -218,6 +226,7 @@ metadata:
 spec:
   auth_service: someservice
   proto: grpc
+  protocol_version: v3
 ---
 apiVersion: getambassador.io/v3alpha1
 kind: Mapping
@@ -228,14 +237,22 @@ spec:
   hostname: "*"
   prefix: /httpbin/
   service: httpbin
+  host_redirect: true
     """
     econf = _get_envoy_config(yaml)
 
+    # check https_redirect variant route
     for rv in econf.route_variants:
         if rv.route.get("match").get("prefix") == "/httpbin/":
             xfp_http_redirect = rv.variants.get("xfp-http-redirect")
             assert xfp_http_redirect
             assert "redirect" in xfp_http_redirect
-            per_filter_config = xfp_http_redirect.get("typed_per_filter_config")
-            assert per_filter_config.get("envoy.filters.http.ext_authz")
-            assert per_filter_config.get("envoy.filters.http.ext_authz").get("disabled") == True
+            _assert_ext_auth_disabled(xfp_http_redirect)
+
+    # check host_redirect route
+    for route in econf.routes:
+        if route.get("match").get("prefix") == "/httpbin/":
+            redirect = route.get("redirect")
+            assert redirect
+            assert redirect.get("host_redirect") == "httpbin"
+            _assert_ext_auth_disabled(route)